THOR-1244: Workaround for v2 PMIC S2 bite
In v2 of the 8941 PMIC, there is a hardware bug. When holding the
power button for ~10 seconds, it is supposed to power down the
device. However, it restarts the device instead.

To workaround this, LK will check the last power off status of PMIC,
if it is due to power button, it will sleep for 5 seconds prior
to actually shutting down the device. This will give users an
impression that holding the power button for 10-15 seconds will shut
off the device.

Change-Id: I38bb2817aafcc0e25e5d692a591300d757c2f6e6
Signed-off-by: Donald Chan <hoiho@lab126.com>