dev: pm8x41: Revise the ldo control apis.
The pmic apis for ldo control assume that the apis work on a
static array defined in the pmic file.
This makes extending these apis to new targets difficult.

Refactored these functions so that the ldo to be worked on
can be passed in from outside of the pmic driver.

CRs-Fixed: 495162
Change-Id: I99b7e9bfa5f74a83f958651b09d43dbddbe5e875