URSA: Display: Use correct clock configuration for splash screen
Previously, the macro used to select which clock configuration was run
wasn't defined.  Coincidentally, the settings the reference device used
worked in most cases until now so the issue wasn't caught sooner.

This change will define CONFIG_MACH_URSA when compiling LK to work on
the Ursa device, for use in temporarily forking code until it can
be refactored to gate on the board_id variable properly

Change-Id: Icb3d39c785caa895c22c91a25b80ffec44ac8c5d