THOR-3240: Reenable mechanism to read msgs from recovery (1/5)
[Problem]
If a system update or factory reset was interrupted by a
shutdown of the device, recovery was not reentering to
finish the process when the device was turned on again,
which could cause the system to be unusable.

[Solution]
AOSP uses a /misc partition to store recovery persistent
information, so it can be read by the bootloader in case there
is a shutdown in the middle of the recovery process.

Reenable mechanism that share messages between recovery
and bootloader through the use of /misc partition.

[Test]
Verified that messages from recovery are properly read by
bootloader. System update and factory reset continued after a
power off/on cycle.

Change-Id: Ia15ea83471f5d870688af8823dbce406b715db88
Signed-off-by: Axel Castaneda Gonzalez <aggonzal@lab126.com>