URSA-603: Display: Enable animation in FBGFX on URSA
The 8974 MDSS requires a trigger to send out a new frame after the
framebuffer has been altered when running in command mode.  This
change adds that "kickoff" trigger to FBGFX's flip routine on URSA

If more platforms with command mode displays are added in the future,
this code should probably be refactored into a platform-specific
callback function.

Change-Id: I4234743e5928168ccc599a80fb7b106070265353