platform: msm_shared: Fix ADMA error
ADMA error is seen with ST_FDS(Fetch descriptor error)
while reading or writing to mmc. This was happening
since the descriptor table allocated for ADMA was not
on the cache line boundaries. Fix descriptor allocation
to fall on cache line boundaries. Also fix the order of
registers writes as per the flow chart in the sdcc spec
version 3.0.

CRs-Fixed: 499156
Change-Id: Ifa571a8b866e250a5b423274205cea64fe23b433