platform: msm_shared: Fix max length for ADMA desc line
Max length for ADMA descriptor line is 65536. As per SD
Spec 3.0 writing '0' to the length field indicates 65536
bytes. Length field is 16 bit unsigned integer and In the
current implementation len is written with value 65536 to
make the code more readable, writing '0' to len might be
confusing. This was done Assuming GCC would truncate 65536
for uint16 to '0'. However we might run into issues if the
toolchain considers to treat the value as
(65536 % max(16 bits unsigned variable)) in case of an
overflow.

CRs-Fixed: 502971
Change-Id: I54229aac0732a82968d3d63f9d7a7965fbb68bf6