URSA-1105: idme: Add ability to update the stored IDME table
Previously we could never update the IDME table stored in emmc once
it had been created.

This implements a simple updater that will add any new fields to the
end of IDME.  It will fail if the existing IDME did not line up with the
default template, or if the IDME version # check fails.

This should normally be disabled, and only enabled for a device if it needs
to update the IDME on existing devices.  Once all desired devices have been
updated, the updater should be turned off.  To enable, define
WITH_IDME_UPDATE_TABLE in the project.mk file.

Change-Id: Ib8783c3b1231ac221605520070051bfd55f936df