APOLLO-2431: [Pentest: LAB-KIN] Unbounded read in bootloader IDME command handler[1/1]
[Problem]
Because the comparison is done using memcmp instead of strncmp, null terminators are ignored on string inputs, and the same length is used for both buffers. Because the length argument is taken from the user-supplied input argv[2], this can read beyond the ends of the version table strings.

[Solution]
use strcmp instead of memcmp.

[Test]
If donot change, input "fastboot oem idme version 2", the version will be  modified to '2'.
After change, input "fastboot oem idme version 2", the version cannot be modified and show "Error version, try again!"

[QA Notes]

Change-Id: Icd1ed9f3d5c88eaba3c357b423fd50c2223248d5