APOLLO-2650: device not boot up battery 2% [1/1]
[Problem]
Device did not boot up with voltage greater than 3.4 V
and capacity greater than 2%

[Solution]
changed the minimal boot up battey capacity from 5% to 2%

[Test]
Tested with Apollo EVT2 device and the device boots up when
the battery capacity reached to 2%.

[QA Notes]
Flash the code with the patch in, drain the battery to 0% and plug
in the charger (SDP or DCP), connect the console log. The device
will boot up when battery capacity reached to 2% (the log will shown
the capacity reached to 2% five times before the device boot up, this
is by design).

Change-Id: Iaafd91f7218a12151e0eb9e746ae787821a0b176
Signed-off-by: Xin Li <xil@lab126.com>