dev: gcdb: Pass default string for unknown dsi panel
If a panel is detected as unknown or if no panel is detected
pass a default string disabling the override of the panel via
command line and disable continuous splash. This will prevent
startup from hanging in the case of continuous splash being enabled
for the default panel in the kernel when no panel is detected
in the bootloader.

Change-Id: I96d69fe56a1bfbd56b8bcf5e42f818147b79ed68