platform: msm_shared: Update the DSI register values
Update the register value of ERR_INT_MASK0 to enable
the ACK_ERR_STATUS only to generate error interrupts.
This avoids unintended flooding of DSI error
interrupts during target boot up in kernel.

Update the POWER_MODE bit of COMMAND_MODE_DMA_CTRL
register to LP mode. This is the default mode that
is set in display drivers in kernel.

Change-Id: Ibbeb0b64068fbbfa525eba974834e04f86c838b2