platform: msm_shared: Add support to use DMA pipe for splash logo
It's possible that some targets that have only a single RGB pipe can have
issues in case of secure video content playback before first suspend resume
if the RGB pipe is used in handoff of splash logo. During continuous splash,
userspace will call prepare on RGB pipe, and as this pipe is being used in
handoff then fallback case happens and VIG pipe is allocated. This leads to
issues in case of secure video content playback as VIG pipe is not available.
To avoid such issues, use DMA pipe instead of RGB for continuous splash handoff.

Change-Id: I8d9d1c2e92c0ab6cf273ead8ebbd5f73a3ac8f23