platform: msm_shared: Fix clock supply to card
When enabling HS400 mode the sdcc clock is run at 400MHZ and is not
reverted back to 200MHZ after enabling hs200 mode, due to this clock
supply to the card in DDR, HS_TIMING mode is double of the value
given by the spec. Revert back to 200MHZ after executing tuning and
disable mci clock divider so that we can use sdhc clock divider to
supply clock to the card.

CRs-Fixed: 703891
Change-Id: Id6e19f2df6ef0139bcfcb104c04f64fddc0c0d64