platform: msm8994: separate core and pixel clocks
Provide different APIs to program core/ahb and pixel clocks
for HDMI. Core clocks are needed for HPD and cable detection.
Once the cable detection is successful, HDMI power on is done
which needs pixel clock. As they are two different code flows,
we need to separate out the pixel clocks.

Change-Id: Id9046f59d4f74b5d5801d9036279bbbca4691dca