platform: msm_shared: use base + offset programming style for DSI
Modify the code to specify register addresses as an offset from the DSI
controller base. This removes the hard dependency on a particular
controller and allows display to be brought up using any DSI controller.
In order to achieve this, separate out all programming for legacy
targets to a single compiler define DISPLAY_TYPE_MDSS and ensure that
all the programming for supported targets use the new style.

Change-Id: Ideb119c3a90cefa1899d15e443af36e2854d316d