platform: msm8960: set boot_done bit
The boot_done flag if not set, causes the PMIC to reset when 90 seconds
have elapsed after power on. This reset feature was introduced to ensure
that the battery was strong enough to bootup the device, in case it was
not the reset will charge the battery before trying the next bootup.
However on Automotive project the PMIC runs on a battery less configuration
so it is safe to set the boot_done bit in the appsbl to prevent the 90
second reset.

Change-Id: I122a1c516928b4a461e70f1e2d5a6258af416458