platform: msm_shared: Fix erased page detection.
On newer nand cards depending on the process technology used there could
be bit flips on pages on the NAND card. On an erased page the controller
fills the page with all 1's. When a read is performed an erased page
because there are bit flips the controller would not detect the page as
erased page instead throws an uncorrectable ecc error. Because of this
an erased page would be considered as a bad page.

To workaround this issue, when driver sees an uncorrectable ECC error
during a page read, it will read the ECC bytes from a page and count
number of zeros in each codeword. If the number of zeros is less than a
threshold value defined by the NAND data sheet then the page can be
safely considered as an erase page.

Change-Id: Ie1dbfb26fae43171a23a816ebf34daeabd634c7f