platform: msm_shared: Update the boot device in hlos subtype
Hlos subtype is updated with boot device in the init of board driver.
For some platforms the fuse address for boot device varies base on soc
version. Boot device driver depends on board driver for soc version.
This creates a circular dependency. In order to avoid this, update the
boot device in hlos subtype after boot device is read.

Change-Id: I1309152bd7bd38a9e3afbf1e71884aafdbe70cce