target: msm9816: Set default XO adjustment
This setting has to do with the capacitance between the XTAL_IN
and XTAL_OUT. With the modem loaded, this value was 0x20.
Without it, it is 0x0. So when the modem image is loaded,
this setting is done by the DSP and everything is ok on Android.
If you don't load the modem image, the setting
(which is needed for WLAN RF to work properly) is not done.

Change-Id: Ie668c4e8e8c123e8256c5550ef35c9a5f96277f3