aboot: set local-mac-address in DTS
Set local-mac-address for the WCNSS node, if not set already. To set it, we rely
on the unique SN provided the target library (typicall the eMMC CID).

The SN is formatted as an 8-char number with leading zero's if needed. The MAC
address is formed by adding '20:00' in front of the SN number to make sure that
we use a MAC address from the locally adminstrated pool.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>
(cherry picked from commit c0026dd80f0cb26a0b32169e66837ea30796f218)