aboot: set wcnss-bt local-mac-address in DTS
Set local-mac-address for the WCNSS BT node, if not set already.
To set it, we rely on the unique SN provided the target library
(typicall the eMMC CID).

The SN is formatted as an 8-char number with leading zero's if needed. The MAC
address is formed by adding '20:00' in front of the SN number to make sure that
we use a MAC address from the locally adminstrated pool.
The implementation is same as for wcnss-wlan MAC address,
but the last bit is flipped to differ between the two addresses.

Change-Id: I3e22ee96cf365aa8a3beee450987628aee835371