app: aboot: Force display_panel to "none"
If Display Splash Screen is enabled, and if the display_panel string in
the device info partition is empty, then set the string to "none". This
makes LK not enable the splash screen by default. It can be enabled
by setting the display_panel using fastboot command:

"fastboot oem select-display-panel <name>"

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>
Signed-off-by: Archit Taneja <architt@codeaurora.org>